+++ src/app/App.jsx
@@ -1,10 +1,13 @@
 import { useState } from 'react'
-// import '/src/assets/stylesheets/App.css'
+import Questionary from '../components/Questionary'
+
+//import '/src/assets/stylesheets/App.css'
 import Game from "/src/components/Game.jsx"
 
 export default function App() {
   return (
     <div className="App">
+      <Questionary />
       <Game />
     </div>
   )
