+++ src/assets/stylesheets/index.css
@@ -1,5 +1,7 @@
 @import "turty.css";
 @import "game.css";
+@import "questionary.css";
+
 
 /* :root {
   font-family: Inter, Avenir, Helvetica, Arial, sans-serif;
