+++ src/assets/stylesheets/turty.css
@@ -2,4 +2,5 @@
   position: absolute;
   bottom: 5%;
   left: 50%;
+  width: 0px;
 }
