+++ src/components/Question.jsx
@@ -1,0 +1,17 @@
+import { useState } from 'react';
+
+
+export default function Question() {
+
+  
+  const [count, setCount] = useState(0);
+
+  return (
+    <div>
+    <p>You clicked {count} times</p>
+      <button onClick={() => setCount(count + 1)}>
+    Click me
+  </button>
+  </div>
+  )
+}
