--- src/questionary/Questionary.jsx
@@ -1,29 +1,0 @@
-import React from 'react'
-import './questionary.css'
-
-const Questionary = () => {
-  return (
-    <div className='container'>
-
-      <div className='question'>
-      </div>
-
-
-      <ul className='answers'>
-        <li className='answer'>Question 1</li>
-        <li className='answer'>Question 2</li>
-        <li className='answer'>Question 3</li>
-        <li className='answer'>Question 4</li>
-      </ul>
-
-      <button className="submit"></button>
-
-
-
-
-    </div>
-
-  )
-}
-
-export default Questionary
